worker/requirements.txt
redis==5.0.8
pymongo==4.8.0
python-dotenv==1.0.1
